Adding all files for Sign block from V1.0.0-beta
Adding the folder containing all of the files used to create the custom block "Pokemart Sign", which was the first thing this mod contained at the time of it's beta release, October 13th, 2025. It contains its textures, 3d models, 3d animation, the block's procedures, as well as screenshots of the block in game.