Adding all files for 2nd Sign block from V1.0.1
Adding the folder containing all of the files used to create the custom block "Pokecenter Sign", which was the 2nd block this mod contained at the time of V1.01's release, October 14th, 2025. It contains its textures, 3d models, and screenshots of the block in game, but no 3d animations or procedures, since the block didn't move at all or do anything really. V1.0.1 still contained the same version of the Pokemart Sign (the first block in the mod), so it was not updated here.