adding the item texture for 1.0.3
So in 1.0.3 I created a usable item texture for the Garbage Can (the texture for the item in the player's hand and in the menus, since it was using the vanilla cauldron's texture in 1.0.2), so I'm adding the new texture to main in order to get ready for 1.0.3's release. I might end up changing it later since it isn't perfect, but it works better for now for sure.