adding unused item textures
So in 1.0.3 I created a usable item texture for the Garbage Can (the texture for the item in the player's hand and in the menus, since it was using the vanilla cauldron's texture in 1.0.2), but I'm adding the UNUSED item textures to main as well in order to get ready for 1.0.3's release. Just want them uploaded to GitHub in-case they are ever needed again for whatever reason.