fixed 1.0.3 changelog's bowl error
So the original changelog uploaded to curseforge said that the trashcan used the vanilla bowl's item texture, but it actually uses the cauldron's item texture. I didn't know I could go back and change the changelogs on curseforge until now, so I've fixed it there and am updating this file to match it.